refactor : 동시성 테스트 공통 부분 통일
